STY: moved F107 support to methods
Moved the local F10.7 routines for additional calculations and internal support to the F10.7 methods sub-module.
